requirements.txt
Django==3.2.16
django-admin-autocomplete-filter==0.7.1
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5
